[Model] Refactor common functionality across Cosmos models into utils file.
